Update Real AS1.R
These were the instructions for this project:

Consider you have been hired to work at the Central Bank of an OECD-member. You are asked to investigate about household indebtedness and suggest policy changes to the institution. Your employer is interested in understanding more about the relationship between unemployment and household debt in their country and needs your expertise to propose possible economic changes. They request you to use the following data from the Federal Reserve Economic Data (FRED) using the pdfetch command:
• Household debt to GDP: ratio, quarterly, not seasonally adjusted
• Unemployment Rate (Aged 15-64 – All Persons): percent, quarterly, seasonally adjusted

They also ask you to present the following results:
1) Address seasonality in the series if necessary.
2) Check if there are trends in each series and address them if there are.
3) Check if the series are stationary and address it if necessary.
4) Find out if there are structural breaks in the household debt series and, if yes, what 
has caused it. Moreover, address it if necessary.
5) Estimate the following model (1): 
𝑑𝑒𝑏𝑡𝑡 = 𝛼0 + 𝛼1𝑑𝑒𝑏𝑡𝑡−1 + 𝛼2𝑢𝑛𝑒𝑚𝑝𝑙𝑜𝑦𝑚𝑒𝑛𝑡𝑡 + 𝜀𝑡 (1),
whose model design is based on previous assessments from the Central Bank. Test if there is a more appropriate model. If yes, present it and compare it to model (1).
6) Check for heteroskedasticity and serial correlation. Address it if necessary.
7) Finally, interpret the results based on the final model you estimated.

I was awarded a First Class Grade in this assignment